# server/docs/swagger.yaml
basePath: /api
definitions:
  department_res.DepartmentResponse:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
  entity.Option:
    properties:
      id:
        type: integer
      name:
        type: string
      tagID:
        type: integer
    type: object
  exam_res.ExamResponse:
    properties:
      filterConditions:
        items:
          $ref: '#/definitions/exam_res.FilterCondition'
        type: array
      id:
        type: string
      questions:
        items:
          $ref: '#/definitions/question_res.QuestionResponse'
        type: array
      semester:
        type: string
      subject:
        type: string
      totalQuestion:
        type: integer
    type: object
  exam_res.FilterCondition:
    properties:
      expectedCount:
        type: integer
      id:
        type: integer
      questions:
        items:
          $ref: '#/definitions/question_res.QuestionResponse'
        type: array
      tagAssignments:
        items:
          $ref: '#/definitions/exam_res.TagAssignment'
        type: array
    type: object
  exam_res.FilterQuestionsList:
    properties:
      expectedCount:
        type: integer
      questions:
        items:
          $ref: '#/definitions/exam_res.QuestionFilterExam'
        type: array
      tagAssignments:
        items:
          $ref: '#/definitions/exam_res.TagAssignment'
        type: array
    type: object
  exam_res.QuestionFilterExam:
    properties:
      answer:
        $ref: '#/definitions/question_res.AnswerResponse'
      content:
        type: string
      id:
        type: string
      isUsed:
        type: boolean
      subQuestions:
        type: object
      tagAssignments:
        items:
          $ref: '#/definitions/question_res.TagAssignmentResponse'
        type: array
      type:
        type: string
    type: object
  exam_res.TagAssignment:
    properties:
      id:
        type: integer
      option:
        $ref: '#/definitions/option_res.OptionResponse'
      tag:
        $ref: '#/definitions/tag_res.TagResponse'
    type: object
  option_res.OptionResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  question_res.AnswerResponse:
    properties:
      content:
        type: object
      id:
        type: string
    type: object
  question_res.OptionResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  question_res.QuestionResponse:
    properties:
      answer:
        $ref: '#/definitions/question_res.AnswerResponse'
      content:
        type: string
      id:
        type: string
      subQuestions:
        type: object
      tagAssignments:
        items:
          $ref: '#/definitions/question_res.TagAssignmentResponse'
        type: array
      type:
        type: string
    type: object
  question_res.TagAssignmentResponse:
    properties:
      id:
        type: integer
      option:
        $ref: '#/definitions/question_res.OptionResponse'
      tag:
        $ref: '#/definitions/question_res.TagResponse'
    type: object
  question_res.TagResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  req.CreateExamRequest:
    properties:
      filterTags:
        items:
          $ref: '#/definitions/req.FilterCondition'
        type: array
      numberQuestion:
        type: integer
      subject:
        type: string
    type: object
  req.CreateOptionRequest:
    properties:
      name:
        type: string
      tagId:
        type: integer
    type: object
  req.CreateQuestionRequest:
    properties:
      answer:
        type: object
      content:
        type: string
      isParent:
        type: boolean
      parentId:
        type: string
      tagAssignments:
        items:
          $ref: '#/definitions/req.CreateTagAssignmentRequest'
        type: array
      type:
        type: string
    required:
    - content
    - type
    type: object
  req.CreateTagAssignmentRequest:
    properties:
      optionId:
        type: integer
      tagId:
        type: integer
    type: object
  req.CreateTagOptionRequest:
    properties:
      name:
        type: string
    type: object
  req.CreateTagRequest:
    properties:
      description:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/req.CreateTagOptionRequest'
        type: array
    type: object
  req.EditQuestionRequest:
    properties:
      answer:
        type: object
      content:
        type: string
      id:
        type: string
      tagAssignments:
        items:
          $ref: '#/definitions/req.UpdateTagAssignmentRequest'
        type: array
      type:
        type: string
    type: object
  req.FilterCondition:
    properties:
      id:
        type: integer
      numberQuestion:
        type: integer
      tagAssignments:
        items:
          $ref: '#/definitions/req.TagAssignment'
        type: array
    type: object
  req.QuestionCompileRequest:
    properties:
      answer:
        type: object
      content:
        type: string
      isParent:
        type: boolean
      subQuestions:
        items:
          type: object
        type: array
    type: object
  req.TagAssignment:
    properties:
      id:
        type: integer
      optionId:
        type: integer
      tagId:
        type: integer
    type: object
  req.UpdateExamRequest:
    properties:
      filterTags:
        items:
          $ref: '#/definitions/req.FilterCondition'
        type: array
      id:
        type: string
      numberQuestion:
        type: integer
      subject:
        type: string
    type: object
  req.UpdateTagAssignmentRequest:
    properties:
      id:
        type: integer
      optionId:
        type: integer
      tagId:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      errorKey:
        type: string
      errorMessage:
        type: string
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  subject_res.SubjectResponse:
    properties:
      code:
        type: string
      department:
        $ref: '#/definitions/department_res.DepartmentResponse'
      id:
        type: string
      name:
        type: string
    type: object
  tag_res.TagResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/option_res.OptionResponse'
        type: array
      subject:
        allOf:
        - $ref: '#/definitions/subject_res.SubjectResponse'
        description: 'TODO: Add table subject -> model subject'
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /exam:
    post:
      consumes:
      - application/json
      description: Create a exam
      parameters:
      - description: CreateExamRequest JSON
        in: body
        name: CreateExamRequest
        required: true
        schema:
          $ref: '#/definitions/req.CreateExamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/exam_res.ExamResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a exam
      tags:
      - Exam
  /exams/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a exam
      parameters:
      - description: Exam Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a exam
      tags:
      - Exam
    get:
      consumes:
      - application/json
      description: Get exam
      parameters:
      - description: Exam Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/exam_res.FilterQuestionsList'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get exam
      tags:
      - Exam
    put:
      consumes:
      - application/json
      description: Update a exam
      parameters:
      - description: UpdateExamRequest JSON
        in: body
        name: UpdateExamRequest
        required: true
        schema:
          $ref: '#/definitions/req.UpdateExamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/exam_res.ExamResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update a exam
      tags:
      - Exam
  /exams/{id}/generate-auto:
    post:
      consumes:
      - application/json
      description: Generate exam auto
      parameters:
      - description: Exam Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/exam_res.ExamResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Generate exam auto
      tags:
      - Exam
  /exams/{id}/get-filter-list:
    get:
      consumes:
      - application/json
      description: Get exam filter questions list
      parameters:
      - description: Exam Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/exam_res.FilterQuestionsList'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get exam filter questions list
      tags:
      - Exam
  /latex-compile:
    post:
      consumes:
      - application/json
      description: Compile Latex to PDF
      parameters:
      - description: QuestionCompileRequest JSON
        in: body
        name: QuestionCompileRequest
        required: true
        schema:
          $ref: '#/definitions/req.QuestionCompileRequest'
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Compile Latex to PDF
      tags:
      - Latex Compile
  /options:
    post:
      consumes:
      - application/json
      description: Create a option
      parameters:
      - description: CreateOptionRequest JSON
        in: body
        name: CreateOptionRequest
        required: true
        schema:
          $ref: '#/definitions/req.CreateOptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.Option'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a option
      tags:
      - Option
  /options/{id}:
    post:
      consumes:
      - application/json
      description: Delete a option
      parameters:
      - description: Id int
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a option
      tags:
      - Option
  /options/{id}/get-used:
    post:
      consumes:
      - application/json
      description: Check option is used or not
      parameters:
      - description: Id int
        in: path
        name: Id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Check option is used or not
      tags:
      - Option
  /questions:
    get:
      consumes:
      - application/json
      description: Get all questions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/question_res.QuestionResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all questions
      tags:
      - Question
    post:
      consumes:
      - application/json
      description: Create a question
      parameters:
      - description: CreateQuestionRequest JSON
        in: body
        name: CreateQuestionRequest
        required: true
        schema:
          $ref: '#/definitions/req.CreateQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/question_res.QuestionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a question
      tags:
      - Question
  /questions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a question
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a question
      tags:
      - Question
    get:
      consumes:
      - application/json
      description: Show a question
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/question_res.QuestionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Show a question
      tags:
      - Question
    put:
      consumes:
      - application/json
      description: Edit a question
      parameters:
      - description: EditQuestionReq JSON
        in: body
        name: EditQuestionRequest
        required: true
        schema:
          $ref: '#/definitions/req.EditQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Edit a question
      tags:
      - Question
  /tags:
    get:
      consumes:
      - application/json
      description: Show all tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/tag_res.TagResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Show all tags
      tags:
      - Tag
    post:
      consumes:
      - application/json
      description: Create a tag
      parameters:
      - description: CreateTagRequest JSON
        in: body
        name: CreateTagRequest
        required: true
        schema:
          $ref: '#/definitions/req.CreateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/tag_res.TagResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a tag
      tags:
      - Tag
  /tags/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a tag
      parameters:
      - description: Id int
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a tag
      tags:
      - Tag
    get:
      consumes:
      - application/json
      description: Show a tag
      parameters:
      - description: Id int
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/tag_res.TagResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Show a tag
      tags:
      - Tag
    put:
      consumes:
      - application/json
      description: Edit a tag
      parameters:
      - description: UpdateTagRequest JSON
        in: body
        name: UpdateTagRequest
        required: true
        schema:
          $ref: '#/definitions/req.CreateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Edit a tag
      tags:
      - Tag
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"